+++ frontend/package.json
@@ -8,7 +8,7 @@
 		"@testing-library/user-event": "^12.8.3",
 		"axios": "^0.21.1",
 		"bootstrap": "^4.6.0",
-		"dotenv": "^8.2.0",
+		"nodemon": "^2.0.9",
 		"react": "^17.0.2",
 		"react-bootstrap": "^1.6.1",
 		"react-dom": "^17.0.2",
@@ -16,7 +16,7 @@
 		"web-vitals": "^1.1.2"
 	},
 	"scripts": {
-		"start": "react-scripts start -r dotenv/config src/index.js",
+		"start": "react-scripts start && nodemon -r dotenv/config src/index.js",
 		"build": "react-scripts build",
 		"test": "react-scripts test",
 		"eject": "react-scripts eject"
@@ -38,5 +38,8 @@
 			"last 1 firefox version",
 			"last 1 safari version"
 		]
+	},
+	"devDependencies": {
+		"dotenv": "^8.2.0"
 	}
 }
