+++ frontend/src/services/moviesServices.js
@@ -1,11 +1,14 @@
 import axios from "axios";
-import { BASE_URL, AUTH_API_CONFIG } from "../";
+const { REACT_APP_BASE_URL, REACT_APP_AUTH_API_CONFIG } = process.env;
 
 export default {
 	getMovies: async function () {
+		console.log(REACT_APP_AUTH_API_CONFIG);
 		try {
-			const response = axios.get(BASE_URL, AUTH_API_CONFIG);
-			return response.data;
+			const response = axios.get(REACT_APP_BASE_URL, {
+				headers: { Authorization: REACT_APP_AUTH_API_CONFIG },
+			});
+			return response;
 		} catch (error) {
 			throw error;
 		}
